Using language feature for java 21

+++ sqrl-flink-lib/sqrl-time/src/main/java/com/datasqrl/time/AbstractEpochToTimestamp.java
@@ -1,13 +1,16 @@
 package com.datasqrl.time;
 
 
-import com.datasqrl.function.FlinkTypeUtil;
 import java.time.Instant;
-import lombok.AllArgsConstructor;
+
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.catalog.DataTypeFactory;
 import org.apache.flink.table.functions.ScalarFunction;
 import org.apache.flink.table.types.inference.TypeInference;
+
+import com.datasqrl.function.FlinkTypeUtil;
+
+import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 public class AbstractEpochToTimestamp extends ScalarFunction {
@@ -17,9 +20,9 @@ public class AbstractEpochToTimestamp extends ScalarFunction {
 
   public Instant eval(Long l) {
     if (isMilli) {
-      return Instant.ofEpochMilli(l.longValue());
+      return Instant.ofEpochMilli(l);
     } else {
-      return Instant.ofEpochSecond(l.longValue());
+      return Instant.ofEpochSecond(l);
     }
   }
